different reorder times for different configurations
created a dictionary to hold these configurations and reorder times. Had
to make a way to turn ion_order into a list then change it to a string
so that the dictionary can use it as a key. Then replaced the variable
reorder time with new dictionary and key var.